Mark old exception utils as deprecated (non-breaking)
